* commit 'dc649462b447951faa05ccae9d3b6b1edc8c42ba': (23 commits)
  Fixes conflicting flags error for '\RTC1' and '\O2' for strict aliasing flag test on Windows. ATK-1065.
  Adds AXOM_DEBUG compiler define to Debug and RelWithDebInfo configurations on windows. ATK-1065.
  Removes logic for outputting a file with the cmake test commands.
  Slight tweak to slic::detail::FalseValue for clang.
  Adds access specifiers to axom_utils timer implementation classes.
  Slight tweak to avoid gcc warning about slic::detail::false_value.
  Improves documentation about the intended usage of the Timer implementation files in axom_utils.
  Moved the false_value type from axom_utils to slic's detail namespace. ATK-1065.
  Adds _USE_MATH_DEFINES to config.hpp for Windows. ATK-1065.
  Bugfix for quest in Visual Studio. ATK-1065.
  Added some additional flags to disable compiler warnings in windows builds.
  Fixes some additional warnings when compiling in release mode. ATK-1065.
  Bugfix for quest in Visual Studio. ATK-1065.
  Fixed errors and warnings in Mint and Primal for Visual Studio. ATK-1065.
  Slam examples working in Visual Studio. ATK-1065.
  Fix for Visual Studio warnings about constant expressions. ATK-1065.
  Slam library and tests running on Windows.  ATK-1065.
  Adds windows support for slic::stacktrace().  ATK-1065.
  Adds Visual Studio compiler flags for disabling warnings. AKT-1065.
  Fixes timer test for windows. ATK-1065.
  ...

+++ src/components/primal/src/spatial_acceleration/BVHTree.hpp
@@ -654,7 +654,7 @@ void BVHTree< T,NDIMS >::percolateDown( int parent, int rChild, int lChild )
 
   // STEP 1: pre-allocated space for object at the children
   const int numObjects   = m_tree[ parent ].ObjectArray.size();
-  const int estChildSize = 0.5 * numObjects;
+  const int estChildSize = static_cast<int>(0.5 * numObjects);
   m_tree[ rChild ].ObjectArray.reserve( estChildSize );
   m_tree[ lChild  ].ObjectArray.reserve( estChildSize );
